Rework `qmk compile` to bypass `Makefile`. Add new --filter option.
